Move LoadLuaLib to CLuaMain.Shared
